moved HeroImage inside Hero component
