+++ src/components/location-finder/Hero/HeroImage.test.js
